+++ src/main/java/de/javanarior/vo/types/speaking/package-info.java
@@ -13,10 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package de.javanarior.vo.types.helper;
-
-import de.javanarior.vo.types.Value;
-
-public interface StringValue extends Value<StringValue, String> {
-
-}
+/**
+ * Speaking API's.
+ */
+package de.javanarior.vo.types.speaking;