+++ src/test/java/de/javanarior/vo/types/BigIntegerWrapperTest.java
@@ -17,7 +17,9 @@ package de.javanarior.vo.types;
 
 import static org.testng.Assert.assertEquals;
 import static org.testng.Assert.assertFalse;
+import static org.testng.Assert.assertNotNull;
 import static org.testng.Assert.assertTrue;
+import static org.testng.Assert.fail;
 
 import java.math.BigInteger;
 
@@ -34,6 +36,16 @@ public class BigIntegerWrapperTest {
     private static final BigIntegerValue TWO = new BigIntegerValue(BigInteger.valueOf(2L));
     private static final BigIntegerValue ANOTHER_ONE = new BigIntegerValue(ONE_VALUE);
     private static final OtherBigIntegerValue OTHER_TYPE_ONE = new OtherBigIntegerValue(ONE_VALUE);
+
+    public void testBigIntegerWrapperNullArgument() {
+        try {
+            new BigIntegerValue(null);
+            fail("Null should not be allowed as value.");
+        } catch (IllegalArgumentException exception) {
+            assertNotNull(exception.getMessage());
+            assertFalse(exception.getMessage().isEmpty());
+        }
+    }
 
     public void testEqualsForDifferentTypes() {
         assertFalse(ONE.equals(OTHER_TYPE_ONE));