+++ src/test/java/de/javanarior/vo/types/helper/OtherStringValueNullObject.java
@@ -17,7 +17,7 @@ package de.javanarior.vo.types.helper;
 
 import de.javanarior.vo.types.NullObject;
 
-public class OtherStringValueNullObject extends NullObject<OtherStringValue> implements OtherStringValue {
+public class OtherStringValueNullObject extends NullObject<OtherStringValue, String> implements OtherStringValue {
 
     public OtherStringValueNullObject() {
         super();